+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 -----
+## [v1.5.1] - 2025-01-29
+
+### Modified
+
+- Allow unlock if no effort but not in charge of the management
+- Do not apply effort field on sightning/follower during merge
+
+-----
 ## [v1.5.0] - 2025-12-12
 
 ### Add